% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltransf.R
\name{ecb}
\alias{ecb}
\title{Eisenthal & Cornish-Bowden}
\usage{
ecb(data, unit_S = 'mM', unit_v = 'au', plot = TRUE)
}
\arguments{
\item{data}{a dataframe where the first column is the independent variable, [S], and the reamining columns (as many as experiment replicates) correspond to the dependent variable, v.}

\item{unit_S}{concentration unit.}

\item{unit_v}{time unit.
#' @param plot logical. If TRUE data are plotted.}
}
\value{
Returns a list with the estimated values of Km and Vm.
}
\description{
Obtains Km and Vm using the Eisenthal & Cornish-Bowden method.
}
\references{
Biochem.J.(1974) 139:715-720 (10.1042/bj1390715)
}
\seealso{
double.rec(), HW(), EH()
}
\author{
Juan Carlos Aledo
}
