% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltransf.R
\name{eh}
\alias{eh}
\title{Eadie-Hofstee Transformation}
\usage{
eh(data, unit_S = 'mM', unit_v = 'au', plot = TRUE)
}
\arguments{
\item{data}{a dataframe where the first column is the independent variable, [S], and the remaining columns (as many as experiment replicates) correspond to the dependent variable, v.}

\item{unit_S}{concentration unit.}

\item{unit_v}{time unit.}

\item{plot}{logical. If TRUE the data and fitted line are plotted.}
}
\value{
A dataframe with the values of the transformed variables is returned. The fitted Km and Vm are given as attributes of this dataframe.
}
\description{
Obtain Km and Vm using the Eadie-Hofstee transformation.
}
\examples{
eh(ONPG[, c(1,2)])
}
\seealso{
lb(), hw(), ecb()
}
