% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Progress.R
\name{fE.progress}
\alias{fE.progress}
\title{Progress Curve for Enzyme-Catalyzed Reaction}
\usage{
fE.progress(data, unit_S = 'mM', unit_t = 'min')
}
\description{
Fits the progress curve of an enzyme-catalyzed reaction.
}
\references{
Biochem Mol Biol Educ.39:117-25 (10.1002/bmb.20479).
}
\seealso{
sEprogress(), int.MM()
}
\author{
Juan Carlos Aledo
}
