% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk.R
\docType{data}
\name{hk}
\alias{hk}
\title{Kinetic data for the phosphorylation of glucose catalyzed by hexokinase.}
\format{
A dataframe with 25 rows (conditions assayed) and 3 columns (variables).
}
\usage{
hk
}
\description{
The variable 'a' is the concentration of ATP-Mg2+ in mM.
The variable 'b' is the glucose concentration in mM.
'rate' is given in arbitrary units.
}
\keyword{datasets}
