% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltransf.R
\name{hw}
\alias{hw}
\title{Hanes-Woolf Transformation}
\usage{
hw(data, unit_S = 'mM', unit_v = 'au', plot = TRUE)
}
\arguments{
\item{data}{a dataframe where the first column is the independent variable, [S], and the reamining columns (as many as experiment replicates) correspond to the dependent variable, v.}

\item{unit_S}{concentration unit.}

\item{unit_v}{time unit.}

\item{plot}{logical. If TRUE the data and fitted line are plotted.}
}
\value{
A dataframe with the values of the transformed variables is returned. The fitted Km and Vm are given as attributes of this dataframe.
}
\description{
Obtains Km and Vm using the Hanes-Woolf transformation.
}
\seealso{
double.rec(), EH(), ECB()
}
\author{
Juan Carlos Aledo
}
