% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance.intMM}
\alias{performance.intMM}
\title{Performance of the Linearized Integrated MM Fitting}
\usage{
performance.intMM(Km, Vm, time, error, sd)
}
\arguments{
\item{Km}{Michaelis contant.}

\item{Vm}{maximal velocity.}

\item{error}{it should be one among c('absolute', 'relative').}

\item{sd}{standard deviation of the error.}

\item{unit_S}{concentration unit.}

\item{unit_t}{time unit.}
}
\value{
Returns a dataframe containing the fitted parameters under different conditions (So/Km ratios).
}
\description{
Evaluates the performance of the int.MM() function when estimating the kinetic parameters.
}
\details{
The time interval must be carefully chosen, otherwise artefacts are possible. When sd is different to 0, then an abosolute error normally distributed is added to the variable St.
}
\seealso{
sE.progress(), int.MM()
}
\author{
Juan Carlos Aledo
}
