% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMM.R
\name{rMM}
\alias{rMM}
\title{Kinetics for An Enzyme of Unknown Parameters}
\usage{
rMM(Km_range = c(1,10), Vm_range = c(10,100), unit_S = 'mM', unit_t = 'au', replicates = 1, error = 'a', sd = 0.1*Vm, outlier = 0, plot = TRUE)
}
\arguments{
\item{Km_range}{range of possible values for Km.}

\item{Vm_range}{range of possible values for Vm.}

\item{unit_S}{concentration unit.}

\item{unit_v}{time unit.}

\item{replicates}{number of replicates for the dependent variable.}

\item{error}{it should be one among c('fixed', 'absolute', 'relative').}

\item{sd}{standard deviation of the error.}

\item{outlier}{an integer, when different from zero, a value of v choseen at random will be affected by outlier-fold its corresponding error.}

\item{plot, }{logical. If TRUE data are plotted.}

\item{Km}{Michaelis contant.}

\item{Vm}{maximal velocity.}
}
\value{
Returns a dataframe where the two first column are the substrate concentration and the velocity without error. The last two columns are the mean velocity and its standard error. In between, when required, the different velocity replicates affected of error are given.
}
\description{
Simulates the kinetics of an michaelian enzyme of unknown parameters.
}
\author{
Juan Carlos Aledo
}
