% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMM.R
\name{sMM}
\alias{sMM}
\title{Kinetics for An Enzyme of Known Parameters}
\usage{
sMM(Km, Vm, unit_S = 'mM', unit_v = 'au', replicates = 3, error = a, sd = 0.05*Vm, outlier = 0, plot = TRUE)
}
\arguments{
\item{Km}{Michaelis contant.}

\item{Vm}{maximal velocity.}

\item{unit_S}{concentration unit.}

\item{unit_v}{time unit.}

\item{replicates}{number of replicates for the dependent variable.}

\item{error}{it should be one among c('fixed', 'absolute', 'relative').}

\item{sd}{standard deviation of the error.}

\item{outlier}{an integer, when different from zero, a value of v choseen at random will be affected by outlier-fold its corresponding error.}

\item{plot, }{logical. If TRUE data are plotted.}
}
\value{
Returns a dataframe where the two first column are the substrate concentration and the velocity without error. The last two columns are the mean velocity and its standard error. In between, when required, the different velocity replicates affected of error are given.
}
\description{
Simulates the kinetics of an michaelian enzyme of unknown parameters.
}
\examples{
sMM(Km = 1, Vm = 1)
}
\seealso{
rMM()
}
\author{
Juan Carlos Aledo
}
