% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcTS.R
\docType{data}
\name{TcTS}
\alias{TcTS}
\title{Kinetic data for Trypanosoma cruzi trans-sialidase.}
\format{
A dataframe with 16 rows (conditions assayed) and 3 columns (variables).
}
\usage{
TcTS
}
\description{
The variable 'a' is the concentration of p-nitrophenyl alpha-sialoside (as donor) in mM.
The variable 'b' is the lactose (as acceptor) concentration in mM.
'rate' is given in mM/min.
Biochemistry 2008, 47, 3507–3512 (https://pubmed.ncbi.nlm.nih.gov/18284211)
}
\keyword{datasets}
